# contracts/openapi/public-api.yaml
openapi: 3.1.0
...
info:
  title: EduTwin Public API
  version: 2.0.0
  license:
    name: Course demonstration use
    identifier: LicenseRef-Course-Demonstration
  description: |
    EduTwin browser-facing contract. MySQL is the system of record. Redis values,
    task state, and server-sent events are recoverable projections. Every traced
    response identifies the accepted answer event, analysis job, preallocated or
    persisted snapshot, data versions, and requested and effective model versions.
    Authentication uses a stateless JWT with one active role and all granted roles.
    Role, password, and account changes advance the token version and invalidate older
    tokens. Logout consists of discarding the JWT in the browser.
servers:
  - url: http://127.0.0.1:8080
    description: Local Compose deployment
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/lms/teacher/courses:
    get:
      tags: [Teacher LMS]
      operationId: listManagedCourses
      summary: List courses managed by the current teacher
      responses:
        '200':
          description: Teacher-owned courses with LMS counts.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LmsCourseList'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Teacher LMS]
      operationId: createManagedCourse
      summary: Create a draft course assigned to the current teacher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseMutationRequest'
      responses:
        '201':
          description: Draft course created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LmsCourse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/lms/teacher/courses/{courseId}:
    parameters:
      - $ref: '#/components/parameters/CourseId'
    put:
      tags: [Teacher LMS]
      operationId: updateManagedCourse
      summary: Update a teacher-owned course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseMutationRequest'
      responses:
        '200':
          description: Course updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LmsCourse'
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/lms/teacher/courses/{courseId}/status:
    parameters:
      - $ref: '#/components/parameters/CourseId'
    put:
      tags: [Teacher LMS]
      operationId: changeManagedCourseStatus
      summary: Publish or archive a teacher-owned course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LmsContentStatusChangeRequest'
      responses:
        '200':
          description: Course status changed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LmsCourse'
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/lms/teacher/courses/{courseId}/status/preview:
    parameters:
      - $ref: '#/components/parameters/CourseId'
    post:
      tags: [Teacher LMS]
      operationId: previewManagedCourseStatus
      summary: Preview affected content, students, and submissions before a course transition
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/LmsContentStatusChangeRequest' } } }
      responses:
        '200': { description: Course lifecycle impact., content: { application/json: { schema: { $ref: '#/components/schemas/LifecycleImpactPreview' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/lms/teacher/courses/{courseId}/sections:
    parameters:
      - $ref: '#/components/parameters/CourseId'
    get:
      tags: [Teacher LMS]
      operationId: getManagedCourseOutline
      summary: Return all draft and published course content
      responses:
        '200':
          description: Teacher course outline.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseOutline'
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [Teacher LMS]
      operationId: createCourseSection
      summary: Add a section to a teacher-owned course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SectionMutationRequest'
      responses:
        '201':
          description: Section created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LmsSection'
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/lms/teacher/sections/{sectionId}/lessons:
    parameters:
      - $ref: '#/components/parameters/SectionId'
    post:
      tags: [Teacher LMS]
      operationId: createSectionLesson
      summary: Add a lesson to a teacher-owned section
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LessonMutationRequest'
      responses:
        '201':
          description: Lesson created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LmsLesson'
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/lms/teacher/sections/{sectionId}:
    parameters:
      - $ref: '#/components/parameters/SectionId'
    put:
      tags: [Teacher LMS]
      operationId: updateManagedSection
      summary: Update a draft section
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/SectionMutationRequest' } } }
      responses:
        '200': { description: Section updated., content: { application/json: { schema: { $ref: '#/components/schemas/LmsSection' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/lms/teacher/sections/{sectionId}/status:
    parameters:
      - $ref: '#/components/parameters/SectionId'
    put:
      tags: [Teacher LMS]
      operationId: changeManagedSectionStatus
      summary: Transition a section lifecycle state
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/LmsContentStatusChangeRequest' } } }
      responses:
        '200': { description: Section status changed., content: { application/json: { schema: { $ref: '#/components/schemas/LmsSection' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/lms/teacher/lessons/{lessonId}:
    parameters:
      - $ref: '#/components/parameters/LessonId'
    put:
      tags: [Teacher LMS]
      operationId: updateManagedLesson
      summary: Update a draft lesson
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/LessonMutationRequest' } } }
      responses:
        '200': { description: Lesson updated., content: { application/json: { schema: { $ref: '#/components/schemas/LmsLesson' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/lms/teacher/lessons/{lessonId}/status:
    parameters:
      - $ref: '#/components/parameters/LessonId'
    put:
      tags: [Teacher LMS]
      operationId: changeManagedLessonStatus
      summary: Transition a lesson lifecycle state
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/LmsContentStatusChangeRequest' } } }
      responses:
        '200': { description: Lesson status changed., content: { application/json: { schema: { $ref: '#/components/schemas/LmsLesson' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/lms/teacher/courses/{courseId}/sections/order:
    parameters:
      - $ref: '#/components/parameters/CourseId'
    put:
      tags: [Teacher LMS]
      operationId: reorderManagedSections
      summary: Replace the ordering of all non-archived sections
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ContentOrderRequest' } } }
      responses:
        '204': { description: Sections reordered. }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/lms/teacher/sections/{sectionId}/lessons/order:
    parameters:
      - $ref: '#/components/parameters/SectionId'
    put:
      tags: [Teacher LMS]
      operationId: reorderManagedLessons
      summary: Replace the ordering of all non-archived lessons
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ContentOrderRequest' } } }
      responses:
        '204': { description: Lessons reordered. }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/lms/teacher/courses/{courseId}/roster:
    parameters:
      - $ref: '#/components/parameters/CourseId'
    get:
      tags: [Teacher LMS]
      operationId: getCourseRoster
      summary: Return enrolled and available students
      responses:
        '200':
          description: Course roster.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RosterList'
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/lms/teacher/courses/{courseId}/roster/{studentId}:
    parameters:
      - $ref: '#/components/parameters/CourseId'
      - $ref: '#/components/parameters/StudentId'
    put:
      tags: [Teacher LMS]
      operationId: updateCourseEnrollment
      summary: Enroll, restore, or withdraw a student and append enrollment history
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnrollmentUpdateRequest'
      responses:
        '204': { description: Enrollment updated. }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/lms/teacher/courses/{courseId}/knowledge-skills:
    parameters:
      - $ref: '#/components/parameters/CourseId'
    get:
      tags: [Teacher LMS]
      operationId: listCourseKnowledgeSkills
      summary: List knowledge skills available for assessment authoring
      responses:
        '200': { description: Course knowledge skills., content: { application/json: { schema: { $ref: '#/components/schemas/KnowledgeSkillList' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/lms/teacher/courses/{courseId}/assessments:
    parameters:
      - $ref: '#/components/parameters/CourseId'
    get:
      tags: [Teacher LMS]
      operationId: listManagedAssessments
      summary: List course assessments and submission counts
      responses:
        '200':
          description: Managed assessments.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessmentList'
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [Teacher LMS]
      operationId: createAssessment
      summary: Create a draft objective assessment with questions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssessmentMutationRequest'
      responses:
        '201':
          description: Assessment created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessmentDetail'
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/lms/teacher/assessments/{assessmentId}/status:
    parameters:
      - $ref: '#/components/parameters/AssessmentId'
    put:
      tags: [Teacher LMS]
      operationId: changeAssessmentStatus
      summary: Publish or archive an assessment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssessmentStatusChangeRequest'
      responses:
        '200':
          description: Assessment status changed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessmentDetail'
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/lms/teacher/assessments/{assessmentId}:
    parameters:
      - $ref: '#/components/parameters/AssessmentId'
    put:
      tags: [Teacher LMS]
      operationId: updateManagedAssessment
      summary: Update a draft assessment and replace its questions
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/AssessmentMutationRequest' } } }
      responses:
        '200': { description: Assessment updated., content: { application/json: { schema: { $ref: '#/components/schemas/AssessmentDetail' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/lms/teacher/assessments/{assessmentId}/extend:
    parameters:
      - $ref: '#/components/parameters/AssessmentId'
    post:
      tags: [Teacher LMS]
      operationId: extendManagedAssessment
      summary: Extend an assessment due date and reopen a closed assessment when applicable
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/AssessmentDueDateRequest' } } }
      responses:
        '200': { description: Assessment extended., content: { application/json: { schema: { $ref: '#/components/schemas/AssessmentDetail' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/lms/teacher/assessments/{assessmentId}/cancel:
    parameters:
      - $ref: '#/components/parameters/AssessmentId'
    post:
      tags: [Teacher LMS]
      operationId: cancelManagedAssessment
      summary: Cancel a published or closed assessment with a reason
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/AssessmentCancelRequest' } } }
      responses:
        '200': { description: Assessment cancelled., content: { application/json: { schema: { $ref: '#/components/schemas/AssessmentDetail' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/lms/teacher/assessments/{assessmentId}/clone:
    parameters:
      - $ref: '#/components/parameters/AssessmentId'
    post:
      tags: [Teacher LMS]
      operationId: cloneManagedAssessment
      summary: Clone an assessment and its questions into a new draft
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/AssessmentCloneRequest' } } }
      responses:
        '201': { description: Assessment clone created., content: { application/json: { schema: { $ref: '#/components/schemas/AssessmentDetail' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/lms/teacher/assessments/{assessmentId}/attempt-requests:
    parameters:
      - $ref: '#/components/parameters/AssessmentId'
    get:
      tags: [Teacher LMS]
      operationId: listManagedAssessmentAttemptRequests
      summary: List student attempt requests for a managed assessment
      responses:
        '200': { description: Attempt requests., content: { application/json: { schema: { $ref: '#/components/schemas/AttemptRequestList' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/lms/teacher/assessments/{assessmentId}/attempt-grants:
    parameters:
      - $ref: '#/components/parameters/AssessmentId'
    post:
      tags: [Teacher LMS]
      operationId: grantManagedAssessmentAttempt
      summary: Proactively approve an extra attempt for one enrolled student
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/AttemptRequestCreate' } } }
      responses:
        '201': { description: Attempt grant created., content: { application/json: { schema: { $ref: '#/components/schemas/AttemptRequest' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/lms/teacher/attempt-requests/{requestId}/decision:
    parameters:
      - { name: requestId, in: path, required: true, schema: { type: string, format: uuid } }
    put:
      tags: [Teacher LMS]
      operationId: decideManagedAssessmentAttemptRequest
      summary: Approve or reject a pending student attempt request
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/AttemptRequestDecision' } } }
      responses:
        '200': { description: Attempt request decided., content: { application/json: { schema: { $ref: '#/components/schemas/AttemptRequest' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/lms/teacher/assessments/{assessmentId}/submissions:
    parameters:
      - $ref: '#/components/parameters/AssessmentId'
    get:
      tags: [Teacher LMS]
      operationId: listAssessmentSubmissions
      summary: List student scores for an assessment
      responses:
        '200':
          description: Assessment scores.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GradeList'
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Stateless JWT containing one role plus userId and authorization-version claims; logout discards the token client-side.
...
  responses:
...
    Unauthorized:
      description: Missing or invalid authentication.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Forbidden:
      description: Authenticated principal lacks the required role or scope.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
  schemas:
...
    LmsCourseList:
      type: object
      additionalProperties: false
      required: [items, total]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/LmsCourse' }
        total: { type: integer, minimum: 0 }
...